feat: added a note for the registration form, and added a note display on the registration details page if there is a note. refactored components on the registration forms for better modularity
