+++ package.json
@@ -43,18 +43,18 @@
     "clean": "rm -rf node_modules .next bun.lock"
   },
   "dependencies": {
-    "@base-ui/react": "^1.4.0",
+    "@base-ui/react": "^1.4.1",
     "@dnd-kit/react": "^0.4.0",
     "@react-pdf/renderer": "^4.5.1",
     "@supabase/ssr": "^0.10.2",
-    "@supabase/supabase-js": "^2.103.3",
-    "@tanstack/react-form": "^1.29.0",
+    "@supabase/supabase-js": "^2.104.0",
+    "@tanstack/react-form": "^1.29.1",
     "@tanstack/react-table": "^8.21.3",
-    "@tiptap/extension-placeholder": "^3.22.3",
-    "@tiptap/react": "^3.22.3",
-    "@tiptap/starter-kit": "^3.22.3",
+    "@tiptap/extension-placeholder": "^3.22.4",
+    "@tiptap/react": "^3.22.4",
+    "@tiptap/starter-kit": "^3.22.4",
     "@types/qrcode": "^1.5.6",
-    "@typescript/native-preview": "^7.0.0-dev.20260417.1",
+    "@typescript/native-preview": "^7.0.0-dev.20260423.1",
     "@yudiel/react-qr-scanner": "^2.5.1",
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
@@ -80,25 +80,25 @@
     "react-medium-image-zoom": "^5.4.3",
     "react-scan": "^0.5.3",
     "react-to-print": "^3.3.0",
-    "resend": "^6.12.0",
+    "resend": "^6.12.2",
     "sanitize-html": "^2.17.3",
     "server-only": "^0.0.1",
-    "shadcn": "^4.3.0",
+    "shadcn": "^4.4.0",
     "sharp": "^0.34.5",
     "skeleton": "^4.15.2",
     "sonner": "^2.0.7",
     "systeminformation": "^5.31.5",
     "tabs": "^0.2.0",
     "tailwind-merge": "^3.5.0",
-    "uuid": "^13.0.0",
-    "write-excel-file": "^3.0.8",
+    "uuid": "^14.0.0",
+    "write-excel-file": "^4.0.2",
     "zod": "^4.3.6",
     "zustand": "^5.0.12"
   },
   "devDependencies": {
     "@biomejs/biome": "2.4.12",
     "@playwright/test": "^1.59.1",
-    "@tailwindcss/postcss": "^4.2.2",
+    "@tailwindcss/postcss": "^4.2.4",
     "@testing-library/jest-dom": "^6.9.1",
     "@testing-library/react": "^16.3.2",
     "@testing-library/user-event": "^14.6.1",
@@ -108,25 +108,23 @@
     "@types/react": "^19.2.14",
     "@types/react-dom": "^19.2.3",
     "@types/sanitize-html": "^2.16.1",
-    "@vitest/coverage-v8": "^4.1.4",
-    "@vitest/ui": "4.1.4",
+    "@vitest/coverage-v8": "^4.1.5",
+    "@vitest/ui": "4.1.5",
     "babel-plugin-react-compiler": "^1.0.0",
-    "bun-types": "^1.3.12",
+    "bun-types": "^1.3.13",
     "dotenv": "^17.4.2",
     "happy-dom": "^20.9.0",
     "hono": "^4.12.14",
     "lefthook": "^2.1.6",
     "playwright-bdd": "^8.5.0",
     "react-email": "6.0.0",
-    "supabase": "^2.92.1",
-    "tailwindcss": "^4.2.2",
+    "supabase": "^2.95.0",
+    "tailwindcss": "^4.2.4",
     "tw-animate-css": "^1.4.0",
     "typescript": "^6.0.3",
-    "vitest": "^4.1.4"
+    "vitest": "^4.1.5"
   },
   "overrides": {
-    "flatted": "^3.4.1",
-    "picomatch": "^2.3.2",
-    "next": "^16.2.3"
+    "uuid": "^14.0.0"
   }
 }
