+++ src/app/registration/[eventId]/_components/forms/step2/Step2.tsx
@@ -1,26 +1,14 @@
 "use client";
 
-import { ArrowLeft, ArrowRight, Plus, Trash2, Users } from "lucide-react";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@/components/ui/alert-dialog";
+import { ArrowLeft, ArrowRight, Plus, Users } from "lucide-react";
 import { Badge } from "@/components/ui/badge";
 import { Button } from "@/components/ui/button";
-import { Card, CardContent } from "@/components/ui/card";
-import { Label } from "@/components/ui/label";
-import { useFieldContext } from "@/hooks/_formHooks";
 import useRegistrationStore from "@/hooks/registration.store";
-import type { StandardRegistrationStep2Schema } from "@/lib/validation/registration/standard";
 import { useRegistrationStep2 } from "../../../_hooks/useRegistrationStep2";
-import RegistrationStepHeader from "../RegistrationStepHeader";
+import RegistrationStepCard from "../_utils/RegistrationStepCard";
+import { RemoveParticipantDialog } from "./RemoveParticipantDialog";
+import { ParticipantFields } from "./Step2ParticipantFields";
+import { RemoveParticipantButton } from "./Step2RemoveParticipantButton";
 
 const MAX_OTHER_PARTICIPANTS = 9;
 
@@ -31,11 +19,6 @@ const EMPTY_REGISTRANT = {
   email: "",
   contactNumber: "",
 } as const;
-
-type SubmitEventLike = {
-  preventDefault: () => void;
-  stopPropagation: () => void;
-};
 
 export default function Step2() {
   const form = useRegistrationStep2();
@@ -51,7 +34,7 @@ export default function Step2() {
     });
   };
 
-  function onNext(e?: SubmitEventLike) {
+  function onNext(e?: React.SubmitEvent) {
     if (e) {
       e.preventDefault();
       e.stopPropagation();
@@ -61,101 +44,12 @@ export default function Step2() {
 
   return (
     <form onSubmit={onNext}>
-      <Card className="w-full overflow-hidden rounded-2xl border-none bg-transparent shadow-none ring-0">
-        <RegistrationStepHeader
-          className="bg-card/5"
-          description="Enter the details of the primary registrant and any additional attendees."
-          Icon={Users}
-          title="Participant Details"
-        />
-
-        <CardContent className="space-y-12 px-0 sm:px-6">
-          <div className="flex flex-col items-start justify-between gap-2 sm:flex-row sm:gap-3">
-            <h3 className="pb-2 font-semibold text-foreground text-lg">
-              Primary Registrant
-            </h3>
-            <Badge className="shrink-0" variant="secondary">
-              <form.Subscribe
-                selector={(s) => s.values.otherParticipants?.length ?? 0}
-              >
-                {(otherParticipantsCount) => (
-                  <>
-                    {otherParticipantsCount + 1} / {MAX_OTHER_PARTICIPANTS + 1}
-                  </>
-                )}
-              </form.Subscribe>
-            </Badge>
-          </div>
-
-          <ParticipantFields form={form} />
-
-          <form.AppField mode="array" name="otherParticipants">
-            {(field) => {
-              const otherParticipantsCount = field.state.value?.length ?? 0;
-              const canAddMore =
-                otherParticipantsCount < MAX_OTHER_PARTICIPANTS;
-
-              return (
-                <div className="space-y-12">
-                  {field.state.value?.map((registrant, idx) => {
-                    const hasData =
-                      registrant.firstName ||
-                      registrant.lastName ||
-                      registrant.email ||
-                      registrant.contactNumber;
-
-                    return (
-                      <div
-                        className="slide-in-from-top-2 fade-in relative animate-in space-y-4 rounded-xl pt-4"
-                        key={registrant.id}
-                      >
-                        <div className="flex items-center justify-between">
-                          <h4 className="font-semibold text-foreground text-lg">
-                            Additional Participant #{idx + 1}
-                          </h4>
-                          {hasData ? (
-                            <RemoveParticipantDialog idx={idx} />
-                          ) : (
-                            <Button
-                              className="h-8 px-2 text-destructive hover:bg-destructive/10 hover:text-destructive"
-                              onClick={() => field.removeValue(idx)}
-                              size="sm"
-                              type="button"
-                              variant="ghost"
-                            >
-                              <Trash2 className="mr-1 h-4 w-4" />
-                              Remove
-                            </Button>
-                          )}
-                        </div>
-                        <ParticipantFields form={form} index={idx} />
-                      </div>
-                    );
-                  })}
-
-                  <Button
-                    className="mt-12 h-12 w-full rounded-xl border-2 border-dashed transition-colors hover:border-primary hover:bg-primary/5 hover:text-primary"
-                    disabled={!canAddMore}
-                    onClick={() =>
-                      field.pushValue({
-                        ...EMPTY_REGISTRANT,
-                        id: crypto.randomUUID(),
-                      })
-                    }
-                    type="button"
-                    variant="outline"
-                  >
-                    <Plus className="mr-2 h-4 w-4" />
-                    {canAddMore
-                      ? "Add Another Participant"
-                      : `Maximum ${MAX_OTHER_PARTICIPANTS + 1} participants reached`}
-                  </Button>
-                </div>
-              );
-            }}
-          </form.AppField>
-
-          <div className="flex flex-col-reverse gap-3 border-border/50 border-t pt-4 sm:flex-row sm:items-center sm:justify-between">
+      <RegistrationStepCard
+        className="w-full overflow-hidden rounded-2xl border-none bg-transparent shadow-none ring-0"
+        contentClassName="space-y-12 px-0 sm:px-6"
+        description="Enter the details of the primary registrant and any additional attendees."
+        footer={
+          <div className="flex w-full flex-col-reverse gap-3 border-border/50 border-t pt-4 sm:flex-row sm:items-center sm:justify-between">
             <Button
               className="w-full rounded-xl sm:w-auto"
               onClick={onBack}
@@ -175,114 +69,86 @@ export default function Step2() {
               <ArrowRight className="ml-2 h-4 w-4" />
             </Button>
           </div>
-        </CardContent>
-      </Card>
+        }
+        headerClassName="bg-card/5"
+        Icon={Users}
+        title="Participant Details"
+      >
+        <div className="flex flex-col items-start justify-between gap-2 sm:flex-row sm:gap-3">
+          <h3 className="pb-2 font-semibold text-foreground text-lg">
+            Primary Registrant
+          </h3>
+          <Badge className="shrink-0" variant="secondary">
+            <form.Subscribe
+              selector={(s) => s.values.otherParticipants?.length ?? 0}
+            >
+              {(otherParticipantsCount) => (
+                <>
+                  {otherParticipantsCount + 1} / {MAX_OTHER_PARTICIPANTS + 1}
+                </>
+              )}
+            </form.Subscribe>
+          </Badge>
+        </div>
+
+        <ParticipantFields form={form} />
+
+        <form.AppField mode="array" name="otherParticipants">
+          {(field) => {
+            const otherParticipantsCount = field.state.value?.length ?? 0;
+            const canAddMore = otherParticipantsCount < MAX_OTHER_PARTICIPANTS;
+
+            return (
+              <div className="space-y-12">
+                {field.state.value?.map((registrant, idx) => {
+                  const hasData =
+                    registrant.firstName ||
+                    registrant.lastName ||
+                    registrant.email ||
+                    registrant.contactNumber;
+
+                  return (
+                    <div
+                      className="slide-in-from-top-2 fade-in relative animate-in space-y-4 rounded-xl pt-4"
+                      key={registrant.id}
+                    >
+                      <div className="flex items-center justify-between">
+                        <h4 className="font-semibold text-foreground text-lg">
+                          Additional Participant #{idx + 1}
+                        </h4>
+                        {hasData ? (
+                          <RemoveParticipantDialog idx={idx} />
+                        ) : (
+                          <RemoveParticipantButton idx={idx} />
+                        )}
+                      </div>
+                      <ParticipantFields form={form} index={idx} />
+                    </div>
+                  );
+                })}
+
+                <Button
+                  className="mt-12 h-12 w-full rounded-xl border-2 border-dashed transition-colors hover:border-primary hover:bg-primary/5 hover:text-primary"
+                  disabled={!canAddMore}
+                  onClick={() =>
+                    field.pushValue({
+                      ...EMPTY_REGISTRANT,
+                      id: crypto.randomUUID(),
+                    })
+                  }
+                  type="button"
+                  variant="outline"
+                >
+                  <Plus className="mr-2 h-4 w-4" />
+                  {canAddMore
+                    ? "Add Another Participant"
+                    : `Maximum ${MAX_OTHER_PARTICIPANTS + 1} participants reached`}
+                </Button>
+              </div>
+            );
+          }}
+        </form.AppField>
+      </RegistrationStepCard>
     </form>
-  );
-}
-
-type Step2FieldName =
-  | `registrant.${keyof StandardRegistrationStep2Schema["registrant"]}`
-  | `otherParticipants[${number}].${keyof NonNullable<StandardRegistrationStep2Schema["otherParticipants"]>[number]}`;
-
-interface ParticipantFieldsProps {
-  form: ReturnType<typeof useRegistrationStep2>;
-  index?: number;
-}
-
-function ParticipantFields({ form, index }: ParticipantFieldsProps) {
-  const prefix =
-    index === undefined ? "registrant" : `otherParticipants[${index}]`;
-
-  return (
-    <div className="space-y-6 sm:space-y-12">
-      <div className="grid gap-6 sm:grid-cols-2">
-        <form.AppField name={`${prefix}.firstName` as Step2FieldName}>
-          {(field) => (
-            <div className="space-y-2">
-              <Label className="text-sm">
-                First Name <span className="text-destructive">*</span>
-              </Label>
-              <field.TextField className="h-11 rounded-xl" placeholder="Juan" />
-            </div>
-          )}
-        </form.AppField>
-        <form.AppField name={`${prefix}.lastName` as Step2FieldName}>
-          {(field) => (
-            <div className="space-y-2">
-              <Label className="text-sm">
-                Last Name <span className="text-destructive">*</span>
-              </Label>
-              <field.TextField
-                className="h-11 rounded-xl"
-                placeholder="Dela Cruz"
-              />
-            </div>
-          )}
-        </form.AppField>
-      </div>
-
-      <div className="grid gap-6 sm:grid-cols-2">
-        <form.AppField name={`${prefix}.email` as Step2FieldName}>
-          {(field) => (
-            <div className="space-y-2">
-              <Label className="text-sm">Email Address</Label>
-              <field.TextField
-                className="h-11 rounded-xl"
-                placeholder="juan.delacruz@example.com"
-                type="email"
-              />
-            </div>
-          )}
-        </form.AppField>
-
-        <form.AppField name={`${prefix}.contactNumber` as Step2FieldName}>
-          {(field) => (
-            <div className="space-y-2">
-              <Label className="text-sm">
-                Contact Number / Telephone Number
-              </Label>
-              <field.TextField
-                className="h-11 rounded-xl"
-                placeholder="09XXXXXXXXX / 0XXXX-XXXX"
-              />
-            </div>
-          )}
-        </form.AppField>
-      </div>
-    </div>
-  );
-}
-
-function RemoveParticipantDialog({ idx }: { idx: number }) {
-  const field = useFieldContext();
-
-  return (
-    <AlertDialog>
-      <AlertDialogTrigger className="h-8 px-2 text-destructive hover:bg-destructive/10 hover:text-destructive">
-        <span className="flex items-center">
-          <Trash2 className="mr-1 h-4 w-4" />
-          Remove
-        </span>
-      </AlertDialogTrigger>
-
-      <AlertDialogContent>
-        <AlertDialogHeader>
-          <AlertDialogTitle>Remove Participant?</AlertDialogTitle>
-          <AlertDialogDescription>
-            Any information entered will be lost. This action cannot be undone.
-          </AlertDialogDescription>
-        </AlertDialogHeader>
-        <AlertDialogFooter>
-          <AlertDialogCancel>Cancel</AlertDialogCancel>
-          <AlertDialogAction
-            className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
-            onClick={() => field.removeValue(idx)}
-          >
-            Remove
-          </AlertDialogAction>
-        </AlertDialogFooter>
-      </AlertDialogContent>
-    </AlertDialog>
   );
 }
