+++ src/app/registration/[eventId]/_components/forms/step3/Step3.tsx
@@ -1,41 +1,13 @@
 "use client";
 
-import { ArrowLeft, ArrowRight, CreditCard, Wallet } from "lucide-react";
+import { ArrowLeft, ArrowRight, CreditCard } from "lucide-react";
 import { Button } from "@/components/ui/button";
-import { Card, CardContent } from "@/components/ui/card";
-import { Label } from "@/components/ui/label";
-import { RadioGroup, RadioGroupItem } from "@/components/ui/radio-group";
 import useRegistrationStore from "@/hooks/registration.store";
-import { cn } from "@/lib/utils";
-import { PaymentMethodEnum } from "@/lib/validation/utils";
 import { useRegistrationStep3 } from "../../../_hooks/useRegistrationStep3";
-import {
-  PaymentProofDropzone,
-  RegistrationPaymentSummary,
-} from "../RegistrationPayment";
-import RegistrationStepHeader from "../RegistrationStepHeader";
-
-const PAYMENT_OPTIONS = [
-  {
-    id: "onsite",
-    value: PaymentMethodEnum.enum.onsite,
-    icon: Wallet,
-    title: "Pay Onsite",
-    description: "Pay via cash or card at the event",
-  },
-  {
-    id: "online",
-    value: PaymentMethodEnum.enum.online,
-    icon: CreditCard,
-    title: "Bank Transfer / Online",
-    description: "Pay now and upload receipt",
-  },
-] as const;
-
-type SubmitEventLike = {
-  preventDefault: () => void;
-  stopPropagation: () => void;
-};
+import RegistrationStepCard from "../_utils/RegistrationStepCard";
+import { RegistrationPaymentSummary } from "../RegistrationPayment";
+import Step3PaymentMethodSection from "./Step3PaymentMethodSection";
+import Step3PaymentProofSection from "./Step3PaymentProofSection";
 
 export default function Step3() {
   const form = useRegistrationStep3();
@@ -59,7 +31,7 @@ export default function Step3() {
     });
   };
 
-  function onNext(e?: SubmitEventLike) {
+  function onNext(e?: React.SubmitEvent) {
     if (e) {
       e.preventDefault();
       e.stopPropagation();
@@ -72,33 +44,12 @@ export default function Step3() {
 
   return (
     <form onSubmit={onNext}>
-      <Card className="w-full overflow-hidden rounded-2xl border-none bg-transparent shadow-none ring-0">
-        <RegistrationStepHeader
-          className="bg-card/5"
-          description="Select your preferred payment method and upload proof if paying online."
-          Icon={CreditCard}
-          title="Payment Information"
-        />
-
-        <CardContent className="space-y-6 px-0 sm:px-6">
-          <RegistrationPaymentSummary
-            baseFee={eventDetails?.registrationFee || 0}
-            participantCount={participantCount}
-            sponsorDiscountPerParticipant={sponsorFeeDeduction}
-            sponsoredBy={sponsoredBy}
-          />
-
-          <PaymentMethodSelection form={form} />
-
-          <form.Subscribe selector={(state) => state.values.paymentMethod}>
-            {(paymentMethod) =>
-              paymentMethod === PaymentMethodEnum.enum.online && (
-                <PaymentProofUpload form={form} />
-              )
-            }
-          </form.Subscribe>
-
-          <div className="flex flex-col-reverse gap-3 border-border/50 border-t pt-4 sm:flex-row sm:items-center sm:justify-between">
+      <RegistrationStepCard
+        className="w-full overflow-hidden rounded-2xl border-none bg-transparent shadow-none ring-0"
+        contentClassName="space-y-6 px-0 sm:px-6"
+        description="Select your preferred payment method and upload proof if paying online."
+        footer={
+          <div className="flex w-full flex-col-reverse gap-3 border-border/50 border-t pt-4 sm:flex-row sm:items-center sm:justify-between">
             <Button
               className="w-full rounded-xl sm:w-auto"
               onClick={onBack}
@@ -118,106 +69,28 @@ export default function Step3() {
               <ArrowRight className="ml-2 h-4 w-4" />
             </Button>
           </div>
-        </CardContent>
-      </Card>
-    </form>
-  );
-}
+        }
+        headerClassName="bg-card/5"
+        Icon={CreditCard}
+        title="Payment Information"
+      >
+        <RegistrationPaymentSummary
+          baseFee={eventDetails?.registrationFee || 0}
+          participantCount={participantCount}
+          sponsorDiscountPerParticipant={sponsorFeeDeduction}
+          sponsoredBy={sponsoredBy}
+        />
 
-interface PaymentMethodSelectionProps {
-  form: ReturnType<typeof useRegistrationStep3>;
-}
+        <Step3PaymentMethodSection form={form} />
 
-function PaymentMethodSelection({ form }: PaymentMethodSelectionProps) {
-  return (
-    <form.AppField
-      listeners={{
-        onChange: () => {
-          if (form.getFieldValue("paymentProof")) {
-            form.resetField("paymentProof");
+        <form.Subscribe selector={(state) => state.values.paymentMethod}>
+          {(paymentMethod) =>
+            paymentMethod === "online" ? (
+              <Step3PaymentProofSection form={form} />
+            ) : null
           }
-        },
-      }}
-      name="paymentMethod"
-    >
-      {(field) => (
-        <div className="space-y-4">
-          <Label className="text-base">Payment Method</Label>
-          <RadioGroup
-            className="grid grid-cols-1 gap-4 sm:grid-cols-2"
-            onValueChange={(value) => {
-              const parsedPaymentMethodValue =
-                PaymentMethodEnum.safeParse(value);
-              if (!parsedPaymentMethodValue.success) {
-                return;
-              }
-
-              field.handleChange(parsedPaymentMethodValue.data);
-            }}
-            value={field.state.value}
-          >
-            {PAYMENT_OPTIONS.map((option) => {
-              const Icon = option.icon;
-
-              return (
-                <div className="flex-1" key={option.id}>
-                  <Label
-                    className={cn(
-                      "flex min-h-30 cursor-pointer flex-col items-center justify-center rounded-xl border-2 border-border bg-transparent p-4 text-center transition-all",
-                      field.state.value === option.value &&
-                        "border-primary bg-primary/5",
-                    )}
-                    htmlFor={option.id}
-                  >
-                    <RadioGroupItem
-                      className="peer sr-only"
-                      id={option.id}
-                      value={option.value}
-                    />
-                    <span
-                      className={cn(
-                        "mb-2 inline-flex h-10 w-10 items-center justify-center rounded-full border border-border bg-muted/40 text-muted-foreground",
-                        field.state.value === option.value &&
-                          "border-primary/30 bg-primary/10 text-primary",
-                      )}
-                    >
-                      <Icon className="h-5 w-5" />
-                    </span>
-                    <span className="font-semibold text-lg">
-                      {option.title}
-                    </span>
-                    <span className="mt-1 text-muted-foreground text-sm">
-                      {option.description}
-                    </span>
-                  </Label>
-                </div>
-              );
-            })}
-          </RadioGroup>
-        </div>
-      )}
-    </form.AppField>
-  );
-}
-
-interface PaymentProofUploadProps {
-  form: ReturnType<typeof useRegistrationStep3>;
-}
-
-function PaymentProofUpload({ form }: PaymentProofUploadProps) {
-  return (
-    <form.AppField name="paymentProof">
-      {(field) => {
-        const selectedFile = field.state.value as File | undefined;
-
-        return (
-          <PaymentProofDropzone
-            errorMessages={field.state.meta.errors}
-            onChange={(file) => field.handleChange(file)}
-            value={selectedFile}
-          />
-        );
-      }}
-    </form.AppField>
+        </form.Subscribe>
+      </RegistrationStepCard>
+    </form>
   );
 }
