+++ src/app/registration/[eventId]/_components/forms/step4/Step4NoteSection.tsx
@@ -1,0 +1,23 @@
+import { Card, CardContent } from "@/components/ui/card";
+import type { useRegistrationStep4 } from "../../../_hooks/useRegistrationStep4";
+
+interface NoteSectionProps {
+  form: ReturnType<typeof useRegistrationStep4>;
+}
+
+export default function NoteSection({ form }: NoteSectionProps) {
+  return (
+    <Card className="rounded-2xl border border-border/50 bg-background">
+      <CardContent className="">
+        <form.AppField name="note">
+          {(field) => (
+            <field.TextareaField
+              description="This field is optional. You may add additional context to your registration. Eg. registering for day 1 only. etc."
+              label="Note"
+            />
+          )}
+        </form.AppField>
+      </CardContent>
+    </Card>
+  );
+}
