+++ src/lib/supabase/db.types.ts
@@ -533,6 +533,7 @@ export type Database = {
           eventId: string;
           identifier: string;
           nonMemberName: string | null;
+          note: string | null;
           numberOfParticipants: number | null;
           paymentMethod: Database["public"]["Enums"]["PaymentMethod"];
           paymentProofStatus: Database["public"]["Enums"]["PaymentProofStatus"];
@@ -545,6 +546,7 @@ export type Database = {
           eventId: string;
           identifier: string;
           nonMemberName?: string | null;
+          note?: string | null;
           numberOfParticipants?: number | null;
           paymentMethod: Database["public"]["Enums"]["PaymentMethod"];
           paymentProofStatus?: Database["public"]["Enums"]["PaymentProofStatus"];
@@ -557,6 +559,7 @@ export type Database = {
           eventId?: string;
           identifier?: string;
           nonMemberName?: string | null;
+          note?: string | null;
           numberOfParticipants?: number | null;
           paymentMethod?: Database["public"]["Enums"]["PaymentMethod"];
           paymentProofStatus?: Database["public"]["Enums"]["PaymentProofStatus"];
@@ -1041,21 +1044,7 @@ export type Database = {
           p_identifier: string;
           p_member_type: string;
           p_non_member_name?: string;
-          p_other_participants?: Json;
-          p_payment_method?: string;
-          p_payment_path?: string;
-          p_registrant?: Json;
-          p_sponsored_registration_id?: string;
-        };
-        Returns: Json;
-      };
-      submit_event_registration_standard: {
-        Args: {
-          p_business_member_id?: string;
-          p_event_id: string;
-          p_identifier: string;
-          p_member_type: string;
-          p_non_member_name?: string;
+          p_note?: string;
           p_other_participants?: Json;
           p_payment_method?: string;
           p_payment_path?: string;
@@ -1094,72 +1083,39 @@ export type Database = {
         };
         Returns: Json;
       };
-      upsert_website_content:
-        | {
-            Args: {
-              p_card_placement?: number;
-              p_entry_key: string;
-              p_icon?: string;
-              p_image_url?: string;
-              p_is_active?: boolean;
-              p_section: Database["public"]["Enums"]["WebsiteContentSection"];
-              p_text_type: Database["public"]["Enums"]["WebsiteContentTextType"];
-              p_text_value?: string;
-            };
-            Returns: {
-              cardPlacement: number | null;
-              createdAt: string;
-              entryKey: string;
-              group: string | null;
-              icon: string | null;
-              id: string;
-              imageUrl: string | null;
-              isActive: boolean;
-              section: Database["public"]["Enums"]["WebsiteContentSection"];
-              textType: Database["public"]["Enums"]["WebsiteContentTextType"];
-              textValue: string | null;
-              updatedAt: string;
-            };
-            SetofOptions: {
-              from: "*";
-              to: "WebsiteContent";
-              isOneToOne: true;
-              isSetofReturn: false;
-            };
-          }
-        | {
-            Args: {
-              p_card_placement?: number;
-              p_entry_key: string;
-              p_group?: string;
-              p_icon?: string;
-              p_image_url?: string;
-              p_is_active?: boolean;
-              p_section: Database["public"]["Enums"]["WebsiteContentSection"];
-              p_text_type: Database["public"]["Enums"]["WebsiteContentTextType"];
-              p_text_value?: string;
-            };
-            Returns: {
-              cardPlacement: number | null;
-              createdAt: string;
-              entryKey: string;
-              group: string | null;
-              icon: string | null;
-              id: string;
-              imageUrl: string | null;
-              isActive: boolean;
-              section: Database["public"]["Enums"]["WebsiteContentSection"];
-              textType: Database["public"]["Enums"]["WebsiteContentTextType"];
-              textValue: string | null;
-              updatedAt: string;
-            };
-            SetofOptions: {
-              from: "*";
-              to: "WebsiteContent";
-              isOneToOne: true;
-              isSetofReturn: false;
-            };
-          };
+      upsert_website_content: {
+        Args: {
+          p_card_placement?: number;
+          p_entry_key: string;
+          p_group?: string;
+          p_icon?: string;
+          p_image_url?: string;
+          p_is_active?: boolean;
+          p_section: Database["public"]["Enums"]["WebsiteContentSection"];
+          p_text_type: Database["public"]["Enums"]["WebsiteContentTextType"];
+          p_text_value?: string;
+        };
+        Returns: {
+          cardPlacement: number | null;
+          createdAt: string;
+          entryKey: string;
+          group: string | null;
+          icon: string | null;
+          id: string;
+          imageUrl: string | null;
+          isActive: boolean;
+          section: Database["public"]["Enums"]["WebsiteContentSection"];
+          textType: Database["public"]["Enums"]["WebsiteContentTextType"];
+          textValue: string | null;
+          updatedAt: string;
+        };
+        SetofOptions: {
+          from: "*";
+          to: "WebsiteContent";
+          isOneToOne: true;
+          isSetofReturn: false;
+        };
+      };
     };
     Enums: {
       ApplicationMemberType: "corporate" | "personal";
@@ -1179,7 +1135,8 @@ export type Database = {
         | "company_thrusts"
         | "board_of_trustees"
         | "secretariat"
-        | "landing_page_benefits";
+        | "landing_page_benefits"
+        | "hero_section";
       WebsiteContentTextType: "Paragraph" | "Title" | "Subtitle";
     };
     CompositeTypes: {
@@ -1367,6 +1324,7 @@ export const Constants = {
         "board_of_trustees",
         "secretariat",
         "landing_page_benefits",
+        "hero_section",
       ],
       WebsiteContentTextType: ["Paragraph", "Title", "Subtitle"],
     },
